requirements/base.txt
pydantic_settings==2.0.3
debugpy==1.8.0
psycopg2-binary==2.9.7
alembic==1.12.1
fastapi[standard]==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
regex==2023.10.3
python-dateutil==2.8.2
pycountry==22.9.6